+++ src/App.tsx
@@ -3,6 +3,7 @@ import HomeScreen from "./screens/HomeScreen";
 import CreateChallengeScreen from "./screens/CreateChallengeScreen";
 import MyChallengesScreen from "./screens/MyChallengesScreen";
 import PlayChallengeScreen from "./screens/PlayChallengeScreen";
+import ShapeChallengeScreen from "./screens/ShapeChallengeScreen";
 import type { Screen } from "./types/GameMode";
 
 export default function App() {
@@ -17,5 +18,7 @@ export default function App() {
       return <MyChallengesScreen onNavigate={setScreen} />;
     case "play":
       return <PlayChallengeScreen challengeId={screen.challengeId} onNavigate={setScreen} />;
+    case "shapeChallenge":
+      return <ShapeChallengeScreen onNavigate={setScreen} />;
   }
 }